+++ HISTORY.md
@@ -5,6 +5,9 @@
 ### Enhancements
 * db_bench: add estimate-table-readers-mem benchmark which prints these stats.
 
+### Bug Fixes
+* unit tests: fix GlobalWriteControllerTest.GlobalAndWBMSetupDelay by waiting for the memtable memory release.
+
 ## Fig v2.5.0 (06/14/2023)
 Based on RocksDB 8.1.1